examples/Inference/det_ray_core_api/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
analytics-python==1.4.0
anyio==3.6.2
apex==0.1
appdirs==1.4.4
argcomplete==3.0.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
azure-core==1.26.3
azure-storage-blob==12.15.0
Babel==2.12.1
backcall==0.2.0
backoff==1.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
bleach==6.0.0
blessed==1.20.0
boto3==1.26.95
botocore==1.29.95
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.1
cmake==3.22.4
colorful==0.5.5
comm==0.1.2
conda==23.1.0
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
determined @ file:///opt/determined/wheels/determined-0.20.2.dev0-py3-none-any.whl
dill==0.3.6
distlib==0.3.6
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
executing==1.2.0
fastapi==0.95.0
fastjsonschema==2.16.3
filelock==3.10.0
flatbuffers==23.3.3
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gpustat==1.0.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
hdfs==2.7.0
horovod==0.24.2
httplib2==0.21.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.11
jsonschema==3.2.0
jupyter-archive==3.3.4
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
lmdb==1.4.0
lomond==0.3.3
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
msgpack==1.0.5
multidict==6.0.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==3.1.0
parso==0.8.3
pathspec==0.11.1
petname==2.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy @ file:///tmp/build/80754af9/pluggy_1648042571233/work
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
Pygments==2.14.0
PyNaCl==1.5.0
pynvml==11.5.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==25.0.2
ray==2.3.1
requests @ file:///croot/requests_1678709721434/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///opt/conda/conda-bld/ruamel.yaml.clib_1641552548372/work
s3transfer==0.6.0
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.26.1
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.8.3
tensorflow-datasets==1.3.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.12.0
termcolor==2.2.0
terminado==0.17.1
texttable==1.6.7
tinycss2==1.2.1
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==1.12.0+cu113
torch-tb-profiler==0.4.1
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1673575502006/work
uvicorn==0.21.1
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
yogadl==0.1.4
zipp==3.15.0
zstandard @ file:///croot/zstandard_1677013143055/work